setup-create-key module
